+++ mobile/www/js/systems/SettingsManager.js
@@ -87,7 +87,9 @@ export class SettingsManager {
                 weapon7: '7',
                 weapon8: '8',
                 scrollWheelSwitch: true,
-                flashlight: 'f'
+                flashlight: 'f',
+                dodge: ' ',
+                cycleThrowable: 'q'
             },
             gamepad: {
                 fire: 7, // RT
@@ -98,7 +100,9 @@ export class SettingsManager {
                 pause: 9, // Start
                 prevWeapon: 4, // LB
                 nextWeapon: 3, // Y
-                melee: 11 // R3
+                melee: 11, // R3
+                dodge: 1, // B
+                cycleThrowable: 13 // D-pad Down
             },
             ui: {
                 controlMode: 'keyboard' // 'keyboard' or 'gamepad' - persisted preference